Adds Strumbar::Instrumentation
Refactored Strumbar to use Strumbar::Instrumentation for loading default
instruments.  Strumbar::Instrumentation#load checks for library constants to
determine if it should load subscriptions for various libraries in the system.

This will be the new endpoint for applicable instruments, and will avoid having
unnecessary dependencies other than what's already in place before it.