Adds options hash to instrument loading
In the old instrumentation block, instruments were able to pass an options hash to them for things like rate
limits.  In order to continue providing this in the new API, an options hash has been added to
`Strumbar::InstrumentList#use`.

When adding an instrument, a hash can be passed and will be send to the object upon loading.  If no hash is
passed, a hash containing Strumbar's default_rate will be passed.

When adding an array of instruments, the hash passed will be applied to ALL instruments passed in the array.